requirements.txt
# Core numerical and data-handling stack
numpy>=1.24
pandas>=1.5
scipy>=1.10
scikit-learn>=1.3

# Feature computation and miscellaneous utilities
requests>=2.31
PyYAML>=6.0
psutil>=5.9

# Machine learning models
lightgbm>=4.0
torch>=2.0

# Reinforcement learning dependencies
stable-baselines3>=2.0
# gymnasium provides the modern Gym API required by stable-baselines3
gymnasium>=0.28
